♻️ Refactor: 사용자 리스트 아이템 컴포넌트 이름 정리
