+++ src/pages/Group/pages/Invite/Invite.tsx
@@ -3,7 +3,7 @@ import * as S from "./Invite.styles";
 import Header from "@/components/Header/Header";
 import Search from "@/components/Search/Search";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import List from "@/components/List/List";
+import UserListItem from "@/components/UserListItem/UserListItem";
 import ToastModal from "@/components/ToastModal/ToastModal";
 import ConfirmModal from "@/components/ConfirmModal/ConfirmModal";
 import PageTitle from "@/components/PageTitle/PageTitle";
@@ -158,7 +158,7 @@ export default function Invite() {
           <S.FriendList>
             <S.Title>친구</S.Title>
             {filteredFriends.map((friend: Friend) => (
-              <List
+              <UserListItem
                 key={friend.id}
                 friend={friend}
                 actionButton={{
